requirements.txt
python-dotenv>=1.0.0
supabase>=2.0.0
groq>=0.4.0
flask>=3.0.0



